+++ src/app/completed/page.tsx
@@ -2,7 +2,7 @@
 //@format
 import RenderHeader from '@/app/_components/Header'
 import { useState, useEffect } from 'react'
-import { Flow, getLocalStorageGlobal, lsFlows, useLocalStorage } from '@/app/_components/lib'
+import { Flow, getLocalStorageGlobal, lsFlows, useLocalStorage } from '@/app/_utils/lib'
 
 /**
  * renders the flow box that displays 3 lines of text (the flow learned)
