+++ src/app/learnmoves/move/learn/page.tsx
@@ -13,7 +13,7 @@ import {
   lsUserLearning,
   updateLocalStorageGlobal,
   useLocalStorage,
-} from '@/app/_components/lib'
+} from '@/app/_utils/lib'
 import { useSearchParams } from 'next/navigation'
 import LoadingFallback from '@/app/_components/LoadingFallback'
